requirements.txt
streamlit
streamlit-folium
folium 
geopandas
matplotlib
scikit-learn
seaborn
pandas
numpy
Pillow
opencv-python
shapely 